Use write! instead of print! to handle broken output pipes
Otherwise rust's implementation for print causes an error message to
be printed out, but the situation is expected in unixy utilities like
these which are meant to be put into pipelines.

That is to say, yes it is an error for the pipe to break, but it's an
error that needs to be handled, just as trying to open a file that
doesn't exist results in an error that needs to be handled.